@@ -1,6 +1,6 @@
 {
   "name": "crowi",
-  "version": "1.6.4",
+  "version": "1.7.0-dev",
   "description": "The simple & powerful Wiki",
   "tags": [
     "wiki",
@@ -24,8 +24,8 @@
     "url": "https://github.com/crowi/crowi.git"
   },
   "engines": {
-    "node": "6.x",
-    "npm": "4.x"
+    "node": "8.x",
+    "npm": "5.x"
   },
   "dependencies": {
     "@slack/client": "~3.13.0",
@@ -37,7 +37,9 @@
     "babel-loader": "~7.1.1",
     "babel-polyfill": "~6.23.0",
     "babel-preset-env": "~1.6.0",
+    "babel-preset-flow": "^6.23.0",
     "babel-preset-react": "~6.24.1",
+    "babel-preset-stage-2": "^6.24.1",
     "basic-auth-connect": "~1.0.0",
     "body-parser": "~1.17.2",
     "bootstrap-sass": "~3.3.6",
@@ -49,7 +51,8 @@
     "consolidate": "~0.14.0",
     "cookie-parser": "~1.4.3",
     "csrf": "~3.0.3",
-    "debug": "~2.6.8",
+    "css-loader": "~0.28.7",
+    "debug": "~3.1.0",
     "del": "~3.0.0",
     "diff": "~3.3.0",
     "diff2html": "~2.3.0",
@@ -60,17 +63,10 @@
     "express": "~4.15.2",
     "express-form": "~0.12.0",
     "express-session": "~1.15.4",
+    "extract-text-webpack-plugin": "~3.0.1",
+    "file-loader": "~1.1.5",
     "font-awesome": "~4.7.0",
     "googleapis": "=12.3.0",
-    "gulp": "~3.9.0",
-    "gulp-concat": "~2.6.0",
-    "gulp-cssmin": "~0.1.7",
-    "gulp-jshint": "~1.12.0",
-    "gulp-rename": "~1.2.2",
-    "gulp-sass": "~3.1.0",
-    "gulp-spawn-mocha": "~2.2.1",
-    "gulp-uglify": "~1.4.2",
-    "gulp-watch": "~4.3.5",
     "highlight.js": "~9.12.0",
     "i18next": "~8.4.3",
     "i18next-express-middleware": "~1.0.2",
@@ -84,11 +80,12 @@
     "marked": "~0.3.5",
     "method-override": "~2.3.1",
     "mkdirp": "~0.5.1",
-    "moment": "~2.18.1",
+    "moment": "^2.22.1",
     "mongoose": "4.7.x",
     "mongoose-paginate": "5.0.x",
     "morgan": "~1.8.2",
     "multer": "~1.3.0",
+    "node-sass": "~4.5.3",
     "nodemailer": "~2.7.0",
     "nodemailer-ses-transport": "~1.5.0",
     "plantuml-encoder": "~1.2.4",
@@ -97,17 +94,22 @@
     "react-dom": "~15.5.0",
     "redis": "~2.6.5",
     "reveal.js": "~3.2.0",
+    "sass-loader": "~6.0.6",
     "socket.io": "~2.0.3",
     "socket.io-client": "~2.0.3",
     "sprintf": "~0.1.5",
+    "style-loader": "~0.19.0",
     "swig": "~1.4.0",
     "vinyl-source-stream": "~1.1.0",
-    "webpack": "~2.2.0",
-    "webpack-manifest-plugin": "~1.1.0",
-    "webpack-stream": "~3.2.0"
+    "webpack": "~3.6.0",
+    "webpack-merge": "~4.1.0",
+    "webpack-stream": "~4.0.0"
   },
   "devDependencies": {
     "chai": "~4.1.0",
+    "concurrently": "~3.5.1",
+    "flow-bin": "^0.65.0",
+    "jshint": "~2.9.5",
     "mocha": "~3.4.2",
     "nodemon": "^1.17.2",
     "proxyquire": "~1.8.0",
@@ -117,11 +119,17 @@
   "license": "MIT",
   "scripts": {
     "start": "node app.js",
-    "dev": "nodemon node app.js",
-    "test": "gulp test",
-    "build": "gulp",
+    "jshint": "jshint lib/**/*.js resource/js/**/*.js",
+    "dev": "concurrently --kill-others \"npm run dev:server\" \"npm run watch\"",
+    "dev:server": "DEBUG=crowi:* NODE_ENV=development nodemon --watch ./lib/ --watch ./locales/ --ext 'js,json' --exec node app.js",
+    "dev:client": "NODE_ENV=development webpack --progress --profile --colors --display-error-details --config tools/webpack.config.js",
+    "flow": "flow",
+    "test": "mocha -r test/bootstrap.js --globals chai --reporter dot test/**/*.test.js",
+    "build": "npm run prod:client",
+    "prod:client": "NODE_ENV=production webpack --bail --progress --colors --display-error-details --config tools/webpack.config.js",
     "webpack": "webpack",
-    "postinstall": "gulp"
+    "watch": "npm run dev:client -- --watch",
+    "postinstall": "npm run build"
   },
   "env": {
     "NODE_ENV": "production"
