@@ -30,7 +30,7 @@ module.exports = function(crowi, app) {
     .use(i18nFsBackend)
     .use(i18nSprintf)
     .init({
-      debug: (crowi.node_env === 'development'),
+      //debug: (crowi.node_env === 'development'),
       fallbackLng: [User.LANG_EN_US],
       whitelist: Object.keys(User.getLanguageLabels()).map((k) => User[k]),
       backend: {
