@@ -12,7 +12,7 @@ module.exports = function(crowi) {
     return hasher.digest('hex');
   }
 
-  attachmentSchema = new mongoose.Schema({
+  const attachmentSchema = new mongoose.Schema({
     page: { type: ObjectId, ref: 'Page', index: true },
     creator: { type: ObjectId, ref: 'User', index: true  },
     filePath: { type: String, required: true },
