@@ -8,7 +8,7 @@ module.exports = function(crowi) {
   ;
 
   // TODO: slack 以外の対応
-  updatePostSchema = new mongoose.Schema({
+  const updatePostSchema = new mongoose.Schema({
     pathPattern: { type: String, required: true },
     patternPrefix:  { type: String, required: true },
     patternPrefix2: { type: String, required: true },
