@@ -11,17 +11,8 @@ module.exports = function(crowi, app, req, locals) {
   };
 
   locals.assets = function(file) {
-    var assetList = crowi.getAssetList();
-    var baseName = file.match(/\/([^\/]+)$/)[1];
-
-    var baseNameWithHash = '';
-    if (assetList[baseName]) {
-      baseNameWithHash = assetList[baseName];
-    } else {
-      return file;
-    }
-
-    return file.replace(baseName, baseNameWithHash);
+    // tmp
+    return file
   };
 
   locals.googleLoginEnabled = function() {
